requirements.txt
streamlit
pypdf
Pillow
exifread
pandas
pdf2image
openai
